担当者を登録、更新する際のバグを修正
