+++ to_do_list/laravel/app/Http/Controllers/TaskListController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class TaskListController extends Controller
+{
+    public function init()
+    {
+        return view('task_list');
+    }
+}
