+++ to_do_list/php/css/new_regist.scss
@@ -60,3 +60,7 @@
   }
 
 }
+
+.error__message{
+  color: red;
+}
